Moved Utilities/XrdAdaptor to FWStorage/XrdAdaptor
