+++ FWCore/Services/plugins/BuildFile.xml
@@ -10,7 +10,7 @@
 <use name="FWCore/Reflection"/>
 <use name="FWCore/Services"/>
 <use name="FWStorage/StorageFactory"/>
-<use name="Utilities/XrdAdaptor"/>
+<use name="FWStorage/XrdAdaptor"/>
 <library file="*.cc" name="FWCoreServicesPlugins">
   <flags EDM_PLUGIN="1"/>
 </library>
