+++ FWStorage/XrdAdaptor/doc/multisource_algorithm_design.md
