+++ FWStorage/XrdAdaptor/src/QualityMetric.cc
