+++ FWStorage/XrdAdaptor/src/QualityMetric.h
