+++ FWStorage/XrdAdaptor/src/XrdStatistics.h
@@ -2,7 +2,7 @@
 #define __XRD_STATISTICS_SERVICE_H_
 
 #include "FWStorage/StorageFactory/interface/IOTypes.h"
-#include "Utilities/XrdAdaptor/interface/XrdStatistics.h"
+#include "FWStorage/XrdAdaptor/interface/XrdStatistics.h"
 #include "FWCore/Utilities/interface/propagate_const.h"
 
 #include <atomic>
